requirements.txt
behave>=1.2.6
selenium>=3.141.0
appium-python-client>=1.0.2
